requirements.txt
boto3==1.4.1
PyYAML==3.12
